+++ resources/views/markdown/changelog.md
@@ -3,6 +3,12 @@
 Last Update: June 2, 2024\
 Travis Michael Cross <a href="mailto:travis@not.tv">travis@not.tv</a>
 
+## v0.9.3.11
+June 3, 2024
+
+* Fix the image upload problem.
+* Make removing image EXIF data optional.
+
 ## v0.9.3.10
 June 2, 2024